+++ examples/apps/web-router/src/store/types/requires-auth.js
@@ -1,0 +1,18 @@
+export const requiresAuth = (type) => ({
+  routeChange(entity, payload, api) {
+    // Does this event concern me?
+    if (payload.route !== entity.type) return
+
+    const user = localStorage.getItem("user")
+    if (!user) {
+      api.notify("navigate", {
+        to: "/login",
+        redirectTo: window.location.pathname,
+        replace: true,
+      })
+      return
+    }
+
+    type.routeChange?.(entity, payload, api)
+  },
+})