+++ examples/apps/web-logo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # lit-todomvc
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@2.5.0
+  - @inglorious/logo@2.0.2
+
 ## 0.1.2
 
 ### Patch Changes