+++ examples/apps/web-table/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lit-todomvc
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@2.5.0
+
 ## 0.1.9
 
 ### Patch Changes