+++ packages/ssx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/ssx
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@2.5.0
+
 ## 0.1.1
 
 ### Patch Changes